Fix hygiene split repo guard drift
