+++ tools/dev/maintainability_allowlist.yml
@@ -5,6 +5,15 @@ files:
     max_lines: 1217
     reason: High-density report projection regression matrix kept together by document-output seam.
 functions:
+  tools/dev/hygiene/docker_ci.py::check_docker_ci_dependency_hygiene:
+    max_lines: 637
+    reason: Docker and CI dependency drift checks now live in their own concern module; keep together until split by package-manager surface.
+  tools/dev/hygiene/runtime_policy.py::check_runtime_policy_drift:
+    max_lines: 323
+    reason: Runtime policy matrix validation now lives in its own concern module; keep together to preserve one policy drift report.
+  tools/dev/hygiene/contract_sync.py::check_contract_sync_entrypoint:
+    max_lines: 281
+    reason: Contract sync and UI validation entrypoint guards now live in their own concern module; keep together for workflow/config drift reporting.
   apps/server/tests/use_cases/run/test_post_analysis_whole_run_diagnosis_ranking.py::test_execute_post_analysis_persists_whole_run_diagnosis_summaries:
     max_lines: 353
     reason: End-to-end whole-run diagnosis persistence scenario kept together to preserve setup/assertion context.
